% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lookup_authors_kwb.R
\name{create_lookup_authors_kwb}
\alias{create_lookup_authors_kwb}
\title{Create Lookup Authors KWB}
\usage{
create_lookup_authors_kwb(orcids = get_kwb_orcids())
}
\arguments{
\item{orcids}{character vector with valid ORCIDs (default: get_kwb_orcids())}
}
\value{
data frame with authors information (as required for R packages)
}
\description{
Create Lookup Authors KWB
}
\examples{
orcids <- get_kwb_orcids()
create_lookup_authors_kwb(orcids)

}
