% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_publications_df_for_orcids.R
\name{create_publications_df_for_orcids}
\alias{create_publications_df_for_orcids}
\title{ORCID: get all publications (using rorcid::works())}
\usage{
create_publications_df_for_orcids(orcids = get_kwb_orcids())
}
\arguments{
\item{orcids}{character vector with valid ORCIDs (default: get_kwb_orcids())}
}
\value{
data.frame with all (public!) ORCID publications for provided ORCIDS
}
\description{
ORCID: get all publications (using rorcid::works())
}
\examples{
orcids <- get_kwb_orcids()
create_publications_df_for_orcids(orcids)

}
