% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_publications_df_for_orcids.R
\name{extract_orcid_from_path}
\alias{extract_orcid_from_path}
\title{Helper function: extract ORCID from path column (returned by rorcid::works())}
\usage{
extract_orcid_from_path(path)
}
\arguments{
\item{path}{content of column "path" (returned by \code{\link[rorcid]{works}})}
}
\value{
character vector with ORCIDS
}
\description{
Helper function: extract ORCID from path column (returned by rorcid::works())
}
