% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_publications_df_for_orcids.R
\name{get_kwb_orcids}
\alias{get_kwb_orcids}
\title{Helper function: known ORCIDs of KWB scientists}
\usage{
get_kwb_orcids()
}
\value{
default ORCIDs of KWB scientists
}
\description{
Helper function: known ORCIDs of KWB scientists
}
\examples{
get_kwb_orcids()

}
