% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_publications_df_for_orcids.R
\name{create_publications_df_for_orcids}
\alias{create_publications_df_for_orcids}
\title{ORCID: get all publications (using rorcid::works())}
\usage{
create_publications_df_for_orcids(orcids = c("0000-0001-5483-4594",
  "0000-0001-9134-2871", "0000-0003-0647-7726"))
}
\arguments{
\item{orcids}{character vector with valid ORCIDs (default: c("0000-0001-5483-4594",
"0000-0001-9134-2871", "0000-0003-0647-7726"))}
}
\value{
data.frame with all (public!) ORCID publications for provided ORCIDS
}
\description{
ORCID: get all publications (using rorcid::works())
}
