% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_publications_df_for_orcids.R
\name{get_author_meta_for_orcids}
\alias{get_author_meta_for_orcids}
\title{ORCID: get author metadata}
\usage{
get_author_meta_for_orcids(orcids = get_kwb_orcids())
}
\arguments{
\item{orcids}{character vector with valid ORCIDs (default: get_kwb_orcids())}
}
\value{
data.frame with valid ORCID metadata for authors
}
\description{
ORCID: get author metadata
}
\examples{
orcids <- get_kwb_orcids()
get_author_meta_for_orcids(orcids)

}
